feat: Separate configs for Discord-related information
