+++ gradle.properties
@@ -1,6 +1,6 @@
 org.gradle.jvmargs=-Xmx2G
 org.gradle.parallel=true
 
-mod.version = 2.0-dev
+mod.version = 2.0-beta
 mod.group = dev.spiritstudios
 mod.id = cantilever
