+++ gradle/libs.versions.toml
@@ -1,27 +1,32 @@
 [versions]
 minotaur = "2.+" # https://plugins.gradle.org/plugin/com.modrinth.minotaur
-shadow = "9.4.0" # https://plugins.gradle.org/plugin/com.gradleup.shadow
+shadow = "9.6.1" # https://plugins.gradle.org/plugin/com.gradleup.shadow
 
-minecraft = "26.1"
+minecraft = "26.1.2"
 
 # https://fabricmc.net/develop/
-fabric_loader = "0.18.4"
-fabric_api = "0.144.3+26.1"
-fabric_loom = "1.15-SNAPSHOT"
+fabric_loader = "0.19.3"
+fabric_api = "0.155.2+26.1.2"
+fabric_loom = "1.17-SNAPSHOT"
 
-kaleido = "0.3.3+1.3.2" # https://repo.sleeping.town/folk/sisby/kaleido-config/
+# Note: Kaleido Config is not version-dependent
+config_kaleido = "0.3.3+1.3.2" # https://repo.sleeping.town/folk/sisby/kaleido-config/
 
-jda = "6.3.2" # https://mvnrepository.com/artifact/net.dv8tion/JDA
-discord-webhooks = "0.8.4" # https://mvnrepository.com/artifact/club.minnced/discord-webhooks
+# Night Config is experimental.
+config_night = "3.9.0"
 
-styled-chat = "V5imDrbM" # https://modrinth.com/mod/styled-chat/versions / this is a Modrinth version ID
+jda = "6.5.0" # https://mvnrepository.com/artifact/net.dv8tion/JDA
+# Note: Discord Webhooks has not been updated since 2023.
+discord_webhooks = "0.8.4" # https://mvnrepository.com/artifact/club.minnced/discord-webhooks
+
+styled_chat = "N8ASEpbM" # https://modrinth.com/mod/styled-chat/versions / this is a Modrinth version ID
 
 # https://maven.nucleoid.xyz/eu/pb4
-pb4-api-predicate = "0.8.1+26.1"
-pb4-api-placeholder = "3.0.0-beta.2+26.1"
-pb4-api-playerdata = "0.10.0+26.1"
+pb4_api_predicate = "0.8.2+26.1"
+pb4_api_placeholder = "3.0.0+26.1"
+pb4_api_playerdata = "0.10.1+26.1"
 
-fabric-permissions = "0.7.0" # https://github.com/lucko/fabric-permissions-api/blob/master/USAGE.md
+fabric_permissions = "0.7.0" # https://github.com/lucko/fabric-permissions-api/blob/master/USAGE.md
 
 [plugins]
 fabric_loom = { id = "net.fabricmc.fabric-loom", version.ref = "fabric_loom" }
@@ -34,13 +39,14 @@ minecraft = { group = "mojang", name = "minecraft", version.ref = "minecraft" }
 fabric_loader = { group = "net.fabricmc", name = "fabric-loader", version.ref = "fabric_loader" }
 fabric_api = { group = "net.fabricmc.fabric-api", name = "fabric-api", version.ref = "fabric_api" }
 
-kaleido = { group = "folk.sisby", name = "kaleido-config", version.ref = "kaleido" }
+config-kaleido = { group = "folk.sisby", name = "kaleido-config", version.ref = "config_kaleido" }
+# config-night = { group = "com.electronwill", name = "night-config:json", version.ref="config_night" }
 
 jda = { group = "net.dv8tion", name = "JDA", version.ref = "jda" }
-discordwebhooks = { group = "club.minnced", name = "discord-webhooks", version.ref = "discord-webhooks" }
+discordwebhooks = { group = "club.minnced", name = "discord-webhooks", version.ref = "discord_webhooks" }
 
-styled-chat = { module = "maven.modrinth:styled-chat", version.ref = "styled-chat" }
-pb4-api-predicate = { module = "eu.pb4:predicate-api", version.ref = "pb4-api-predicate" }
-pb4-api-placeholder = { module = "eu.pb4:placeholder-api", version.ref = "pb4-api-placeholder" }
-pb4-api-playerdata = { module = "eu.pb4:player-data-api", version.ref = "pb4-api-playerdata" }
-fabric-permissions = { module = "me.lucko:fabric-permissions-api", version.ref = "fabric-permissions" }
+styled_chat = { module = "maven.modrinth:styled-chat", version.ref = "styled_chat" }
+pb4_api_predicate = { module = "eu.pb4:predicate-api", version.ref = "pb4_api_predicate" }
+pb4_api_placeholder = { module = "eu.pb4:placeholder-api", version.ref = "pb4_api_placeholder" }
+pb4_api_playerdata = { module = "eu.pb4:player-data-api", version.ref = "pb4_api_playerdata" }
+fabric_permissions = { module = "me.lucko:fabric-permissions-api", version.ref = "fabric_permissions" }
